move get_runtime_env_info and parse_runtime_env to ray.utils
