+++ python/ray/actor.py
@@ -5,7 +5,7 @@ import weakref
 import ray.ray_constants as ray_constants
 import ray._raylet
 import ray._private.signature as signature
-from ray.runtime_env import get_runtime_env_info, parse_runtime_env
+from ray.utils import get_runtime_env_info, parse_runtime_env
 import ray.worker
 from ray.util.annotations import PublicAPI
 from ray.util.placement_group import configure_placement_group_based_on_context
