+++ python/ray/job_config.py
@@ -100,7 +100,7 @@ class JobConfig:
         # TODO(edoakes): this is really unfortunate, but JobConfig is imported
         # all over the place so this causes circular imports. We should remove
         # this dependency and pass in a validated runtime_env instead.
-        from ray.runtime_env import get_runtime_env_info
+        from ray.utils import get_runtime_env_info
 
         if self._cached_pb is None:
             pb = gcs_utils.JobConfig()
